# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: PriceHunt API
  description: API para la plataforma de comercio algorítmico PriceHunt
  version: 1.0.0
  contact:
    email: api@pricehunt.mx
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.pricehunt.mx/v1
    description: Production
...
paths:
...
  /products/{id}:
    get:
      tags: [Products]
      summary: Obtener producto por ID
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{id}/variants:
    get:
      tags: [Products]
      summary: Obtener variantes de un producto
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Lista de variantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductVariant'
  /products/{id}/images:
    get:
      tags: [Products]
      summary: Obtener imágenes de un producto
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Lista de imágenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductImage'
...
  /checkout:
    post:
      tags: [Checkout]
      summary: Iniciar proceso de checkout
      description: |
        Valida stock, precios y crea la orden pendiente de pago.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout iniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflicto (precio cambió, sin stock)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutConflict'
  /checkout/{checkout_token}/confirm:
    post:
      tags: [Checkout]
      summary: Confirmar checkout con pago
      parameters:
        - name: checkout_token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Pago procesado, orden creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderConfirmation'
        '402':
          description: Pago fallido
        '409':
          description: Orden ya procesada (idempotency)
  /orders:
    get:
      tags: [Orders]
      summary: Listar órdenes del usuario
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de órdenes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Obtener orden por ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancelar orden
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Orden cancelada
        '400':
          description: No se puede cancelar (ya enviada)
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/tracking:
    get:
      tags: [Orders]
      summary: Obtener tracking de una orden
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Información de tracking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingInfo'
...
  /admin/orders:
    get:
      tags: [Admin]
      summary: Listar todas las órdenes (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de órdenes
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID del producto
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID de la orden
...
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        canonical_name:
          type: string
        slug:
          type: string
        description:
          type: string
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          type: string
        gtin:
          type: string
        mpn:
          type: string
        attributes:
          type: object
          additionalProperties: true
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        best_offer:
          $ref: '#/components/schemas/BestOffer'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        attributes:
          type: object
          additionalProperties: true
        price:
          $ref: '#/components/schemas/Money'
        in_stock:
          type: boolean
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt_text:
          type: string
        position:
          type: integer
        is_primary:
          type: boolean
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    Money:
      type: object
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
          example: MXN
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          default: MX
...
    BestOffer:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        offer_score:
          type: number
        price_score:
          type: number
        delivery_score:
          type: number
        reliability_score:
          type: number
        match_confidence:
          type: number
        supplier:
          $ref: '#/components/schemas/Supplier'
        our_price:
          $ref: '#/components/schemas/Money'
        original_price:
          $ref: '#/components/schemas/Money'
        savings:
          $ref: '#/components/schemas/Money'
        savings_percentage:
          type: number
        shipping_cost:
          $ref: '#/components/schemas/Money'
        total_landed_cost:
          $ref: '#/components/schemas/Money'
        delivery_days:
          type: integer
        in_stock:
          type: boolean
...
    CheckoutRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckoutItem'
          minItems: 1
        shipping_address:
          $ref: '#/components/schemas/Address'
        email:
          type: string
          format: email
        phone:
          type: string
        create_account:
          type: boolean
          default: false
        coupon_code:
          type: string
      required:
        - items
        - shipping_address
        - email
    CheckoutItem:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        supplier_product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - product_id
        - variant_id
        - supplier_product_id
        - quantity
    CheckoutResponse:
      type: object
      properties:
        checkout_token:
          type: string
          description: Token idempotente para este checkout
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckoutItemSummary'
        subtotal:
          $ref: '#/components/schemas/Money'
        shipping:
          $ref: '#/components/schemas/Money'
        tax:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        payment_methods:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [card, oxxo, spei]
              name:
                type: string
        expires_at:
          type: string
          format: date-time
          description: Checkout expira en 15 minutos
    CheckoutItemSummary:
      type: object
      properties:
        product_name:
          type: string
        variant_name:
          type: string
        quantity:
          type: integer
        unit_price:
          $ref: '#/components/schemas/Money'
        total_price:
          $ref: '#/components/schemas/Money'
    CheckoutConflict:
      type: object
      properties:
        type:
          type: string
          enum: [price_changed, out_of_stock, supplier_unavailable]
        message:
          type: string
        item_index:
          type: integer
        old_price:
          $ref: '#/components/schemas/Money'
        new_price:
          $ref: '#/components/schemas/Money'
    PaymentRequest:
      type: object
      properties:
        payment_method:
          type: string
          enum: [card, oxxo, spei]
        card_token:
          type: string
          description: Token de tarjeta (no datos sensibles)
        billing_address:
          $ref: '#/components/schemas/Address'
      required:
        - payment_method
    OrderConfirmation:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [confirmed, processing]
        total:
          $ref: '#/components/schemas/Money'
        estimated_delivery:
          type: string
          format: date-time
        tracking_url:
          type: string
          format: uri
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          $ref: '#/components/schemas/Money'
        shipping:
          $ref: '#/components/schemas/Money'
        tax:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        shipping_address:
          $ref: '#/components/schemas/Address'
        payment_method:
          type: string
        tracking:
          $ref: '#/components/schemas/TrackingInfo'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
        variant_name:
          type: string
        quantity:
          type: integer
        unit_price:
          $ref: '#/components/schemas/Money'
        total_price:
          $ref: '#/components/schemas/Money'
        supplier:
          type: string
        status:
          type: string
    OrderList:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    TrackingInfo:
      type: object
      properties:
        carrier:
          type: string
        service:
          type: string
        tracking_number:
          type: string
        status:
          type: string
        estimated_delivery:
          type: string
          format: date-time
        events:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              location:
                type: string
              timestamp:
                type: string
                format: date-time
              description:
                type: string
    Supplier:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [api, feed, scraping, direct]
        is_active:
          type: boolean
        reliability_score:
          type: number
          minimum: 0
          maximum: 1
        products_count:
          type: integer
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT del login